implemented templates rendering
