+++ index.html
@@ -33,6 +33,8 @@
     </style>
 </head>
 <body>
-<div id="sqlqb"></div>
+<div id="sqlqb">
+    <div id="sqlqb-input-container"></div>
+</div>
 </body>
 </html>
